MNT: re-add examples as separate files
